refactor: enable Composable Transports (#141)
This PR decouples client dialing from Shadowsocks by introducing the StreamDialer and PacketListener types. Those types stand alone, so you are not forced to conflate Stream and Packet handling in a single Client object like we have now.

This PR decouples the Shadowsocks StreamDialer and PacketListener from TCP and UDP by introducing the StreamEndpoint and PacketEndpoint abstractions. This allows one to use Shadowsocks over any transport, including Unix sockets, Websockets, plain TLS, nested Shadowsocks, etc.

The new APIs make use of context.Context, to allow for deadlines or cancellations that were not possible before. We can now impose a deadline on establishing TCP connections in the connectivity tests, which wasn't possible before, delaying the response to the user when traffic was dropped or the server was unresponsive.