refactor: clean up metrics (#144)
This code separates metrics that are generic for proxying from Shadowsocks-specific metrics. This moves towards a place where we can replace the protocol more easily. For TCP, all we need is to replace the handleConnection method. It will also help per-AS metrics.

In the process, I introduced a CountryCode type to clarify when we are using country codes.